Added inspectTransitions function, refactored conditions evaluation (#73)
API changes
====
API remains the same as before, but internals work differently: `findAvailableTransitions` delegates to `inspectTransitions`. It acts similar to previous `findAvailableTransitions` but produces a verbose output which can be then filtered in many ways.

`inspectTransitions` is a low-level API and is mainly meant for debugging purposes.

Shape of output: 

```
[
  {
    transition: <transition as it is defined in schema>,
    result: {
      guards: [
        {
          condition: <guard as defined in schema>,
          result: <evaluation result, boolean>
        },
        ...
      ],
      automatic: <boolean if defined as boolean> OR [
        {
          condition: <automatic as defined in schema>,
          result: <evaluation result, boolean>
        },
        ...
      ],
    }
  }
]
```

So, from this point you can find out why certain transition is not available, etc.